Refactored some code again, re-organized.
ChannelMessage is now a property-based class, making it far more
flexible than it was previously. Enums are set up for IRC protocol
responses and IRC modes are now able to be translated (though I have
very few documented). This current build does not accept commands; will
be working through the night to get it working.